// types.ts
export interface DashboardData {
  row_number: number;
  categoria: string;
  "subcategoria ampla": string;
  "subcategoria especifica": string;
  responsavel: string;
  "competencia / referencia": string;
  valor: number | string;
  indicador: string;
  cidade: string;
  ano: number;
  mes: string;
  data: string;
  status: string;
}